server/requirements.txt
Flask==2.2.2
Flask_SocketIO==5.3.2
pyTelegramBotAPI==4.8.0
python-dotenv==0.21.0
